Updates to view/change feature system
This function was previously not working at all, and now it seems as
though you can switch between different feature systems, although it can
take a few seconds if the corpus is large. There's still a problem where
certain feature files cannot be read properly. Right now any file with
the string 'spe' in its name has to be comma-separated and anything with
'hayes' in the name has to be tab-separated. For reasons unclear to me,
other files, even if formatted identically, will not work.