Changes to PROD
Predictability of distribution now has a "print" button inside of the
results window, making it more like the functional load results. The
results are also now properly formatted as a tab-separated file using
the heading suggested by Kathleen. At this point however it can only do
a single PROD calculation at a time - there's not yet an option to add
more results to the table.