Improved InventoryWindow
IPA-style charts now work for segment selection. Only Frequency of
Alternation has this implemented right now, and there are a few issues
still unresolved. In particular, this only works with SPE features right
now, because it knows to use [voc] to tell apart Cs and Vs. This isn't a
feature in every feature system, so there needs to be a more general
solution for this.