Fixed bug in feature edit window
It was possible for user to select a combination of transcription and
feature system that didn't exist. For example, user has ipa2spe,
ipa2hayes, and arpabet2spe. The option for using arpabet and hayes will
appear, but selecting this caused problems.
Now users are warned about this, informed about the option to download
new feature system, and they are also given the option to create a
feature system full of default values.