Updates to autocategorization of Inventory
Works with Hayes for newly created corpora. Works *most* of the time for
SPE, but for some weird reason the consonant manners of articulation
(only) won't load. If you manually create exactly the same row, it works
fine, so it's just the autocategorization code that is failing.

The uncategorized segments now have a tooltip that tells you any partial
matches in the table.