Changes to feature systems
- You can no longer choose transcription and features independently.
This simplifies the coding, and makes it easier to the user to find the
right system.
- If you try to load a corpus from scratch with symbols that don't match
your feature file, you will be given the option to load with all missing
symbols assigned a default value of 'n' for every feature
- If you swap in a new feature system after loading a corpus, and your
corpus has symbols not found in the new system, then you have to use the
Add Segment button in the Feature menu
- Various other small bugs are also fixed
- There is known bug with the SPE system involving the features [EXTRA]
and [LONG], which are in capital letters. This seems to occasionally
raise a KeyError, but I'm not sure why it doesn't happen all the time.