Fix for issue #359
It is not possible to do long-distance searches using the environment
widget. In addition to "add segments" and "add features", there is now
an option for "match anything here" which puts a * into the environment.
This will, of course, match any segment.  For example, #p*[+voc] matches
word-intial /p/ followed by anything followed by a vowel.

This only seems to work for phonological search right now. Next thing to
do is to fix this for the analysis functions too.