Changes to inventory chart
The inventory chart inside of "Corpus Summary" is now more stretchy, and
should move to fill the window if resized. This gives more space to
display all the segment information.

The uncategorized segment chart was not properly displaying all of the
relevant segments (e.g issue #641). It has been changed from a
QSortFilterProxyModel to a QAbstractItemModel and this seems to have
resolved the problem.