Update for issue #642
Environment positions now have a "allow zero-match" options, which
matches nothing in that position. Suppose you search for {t} in the
middle position, then add {w} and {-consonantal} positions to the right.
This searches for [tw] followed by a vowel. If you set the {w} position
to "allow zero matches", then it will search for words that match
[tw]+vowel or [t]+vowel.