Revised Functional Load script
Added a revised version of the functional load script. This version
lives in a higher-level folder and integrates a number of printed
tests, so it is not yet a *replacement* for the current functional load
script, but it is intended to serve as the base code for such a
replacement. The key difference is that it allows for two different
kinds of relativization within the minimal pairs algorithm.