When the user adds a new word with duplicate spelling and transcription (issue #717)
If both transcription and spelling match an existing entry,
this provides two options: 'add it as a separate entry (with the same
transcription and spelling' or 'add the frequency to the existing entry.'

The user can change the name of a phonological search from the 'load
search' window. Also, this commit enabled the user to save a search from
the 'current searches' table.